backend/pyproject.toml
[tool.poetry]
name = "zus-backend"
version = "0.1.0"
description = "FastAPI backend for the ZUS AI Assistant"
authors = ["Mindhive Candidate <candidate@example.com>"]
readme = "README.md"
packages = [{ include = "backend", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
aiosqlite = "^0.19.0"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
ruff = "^0.3.4"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = ["B008"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
